apis/requirements.txt
minio==7.2.3
PyYAML==6.0.2
python-dotenv==1.0.1
fastapi==0.115.4
uvicorn==0.32.0
SQLAlchemy==2.0.36
psycopg2-binary==2.9.10
pandas==2.2.3
SQLAlchemy-Utils==0.41.2
python-multipart==0.0.17
httpx==0.27.2
pytest==8.3.3
pytest-order==1.3.0
requests==2.32.3